refactor(schemas): migrate BatchVectorizeRequest and ReindexWithContextRequest from inline to knowledge/schemas/ (#5528)
